requirements.txt
python-dotenv~=1.0.0
pyTelegramBotAPI~=4.13.0
psycopg2-binary~=2.9.7